python-ai-service/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Gemini
google-generativeai>=0.3.2

# HTTP client
httpx>=0.26.0

# Redis for caching
redis>=5.0.1

# Environment variables
python-dotenv>=1.0.0

# Data validation (use latest with pre-built wheels)
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
anyio>=4.2.0

# Logging
structlog>=24.1.0
